--- src/core/benchmark/BUILD
@@ -1,5 +1,0 @@
-go_benchmark(
-    name = "benchmark",
-    srcs = ["benchmark_test.go"],
-    deps = ["//src/core"],
-)
