+++ src/core/utils_benchmark_test.go
@@ -1,30 +1,29 @@
-package benchmark
+package core
 
 import (
 	"fmt"
-	"github.com/thought-machine/please/src/core"
 	"testing"
 )
 
 func BenchmarkIterInputsControl(b *testing.B) {
-	state := core.NewDefaultBuildState()
-	target := core.NewBuildTarget(core.NewBuildLabel("src/foo", "foo_lib"))
+	state := NewDefaultBuildState()
+	target := NewBuildTarget(NewBuildLabel("src/foo", "foo_lib"))
 	state.Graph.AddTarget(target)
 
 	b.ReportAllocs()
 
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		for range core.IterInputs(state.Graph, target, true, false) {}
+		for range IterInputs(state.Graph, target, true, false) {}
 	}
 }
 
 func BenchmarkIterInputsSimple(b *testing.B) {
-	state := core.NewDefaultBuildState()
+	state := NewDefaultBuildState()
 
-	target := core.NewBuildTarget(core.NewBuildLabel("src/foo", "foo_lib"))
+	target := NewBuildTarget(NewBuildLabel("src/foo", "foo_lib"))
 	for i := 0; i < 100; i++ {
-		dep := core.NewBuildTarget(core.NewBuildLabel("src/bar", fmt.Sprintf("name_%d", i)))
+		dep := NewBuildTarget(NewBuildLabel("src/bar", fmt.Sprintf("name_%d", i)))
 		for j := 0; j < 200; j++ {
 			dep.AddOutput(fmt.Sprintf("name_%v_file_%v", i, j))
 		}
@@ -33,7 +32,7 @@ func BenchmarkIterInputsSimple(b *testing.B) {
 	}
 
 	for i := 0; i < 25; i++ {
-		target.AddSource(core.FileLabel{
+		target.AddSource(FileLabel{
 			File:    fmt.Sprintf("src_%v", i),
 			Package: "src/foo",
 		})
@@ -44,17 +43,17 @@ func BenchmarkIterInputsSimple(b *testing.B) {
 
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		for range core.IterInputs(state.Graph, target, true, false) {}
+		for range IterInputs(state.Graph, target, true, false) {}
 	}
 }
 
 
 func BenchmarkIterInputsNamedSources(b *testing.B) {
-	state := core.NewDefaultBuildState()
+	state := NewDefaultBuildState()
 
-	target := core.NewBuildTarget(core.NewBuildLabel("src/foo", "foo_lib"))
+	target := NewBuildTarget(NewBuildLabel("src/foo", "foo_lib"))
 	for i := 0; i < 100; i++ {
-		dep := core.NewBuildTarget(core.NewBuildLabel("src/bar", fmt.Sprintf("name_%d", i)))
+		dep := NewBuildTarget(NewBuildLabel("src/bar", fmt.Sprintf("name_%d", i)))
 		for j := 0; j < 200; j++ {
 			dep.AddOutput(fmt.Sprintf("name_%v_file_%v", i, j))
 		}
@@ -64,7 +63,7 @@ func BenchmarkIterInputsNamedSources(b *testing.B) {
 
 	for i := 0; i < 5; i++ {
 		for j := 0; j < 5; j++ {
-			target.AddNamedSource(fmt.Sprintf("srcs_%v", i), core.FileLabel{
+			target.AddNamedSource(fmt.Sprintf("srcs_%v", i), FileLabel{
 				File:    fmt.Sprintf("src_%v_%v", i, j),
 				Package: "src/foo",
 			})
@@ -76,6 +75,6 @@ func BenchmarkIterInputsNamedSources(b *testing.B) {
 
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		for range core.IterInputs(state.Graph, target, true, false) {}
+		for range IterInputs(state.Graph, target, true, false) {}
 	}
 }
